% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddClusterButton.R
\name{AddClusterButton}
\alias{AddClusterButton}
\title{Add Cluster Button}
\usage{
AddClusterButton(map, clusterId)
}
\arguments{
\item{map}{'\link[leaflet]{leaflet}'.
Map widget object}

\item{clusterId}{'character'.
Identification for the marker cluster layer.}
}
\value{
Used for the side-effect of a button placed on the web map.
}
\description{
This function adds a button on a web map that toggles marker clusters
between frozen and unfrozen states.
}
\examples{
map <- CreateWebMap()
d <- maps::us.cities
opt <- leaflet::markerClusterOptions(showCoverageOnHover = FALSE)
id <- "cities_cluster"
map <- leaflet::addMarkers(map, lng = d$long, lat = d$lat, popup = d$name,
                           clusterOptions = opt, clusterId = id)
map <- AddClusterButton(map, id)
map

}
\seealso{
\code{\link{CreateWebMap}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
