% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddHomeButton.R
\name{AddHomeButton}
\alias{AddHomeButton}
\alias{AddClusterButton}
\title{Add Miscellaneous Web Map Buttons}
\usage{
AddHomeButton(map, extent, position = "topleft")

AddClusterButton(map, clusterId, position = "topleft")
}
\arguments{
\item{map}{'\link[leaflet]{leaflet}'.
Map widget object}

\item{extent}{'Raster*', 'Spatial*', 'Extent', 'matrix', or 'numeric'.
Extent object (or object from which an \code{\link[raster]{extent}} object can be extracted/created)
representing a rectangular geographical area on the map.}

\item{position}{'character'.
Position of the button on the web map.
Possible values are \code{"topleft"}, \code{"topright"}, \code{"bottomleft"}, and \code{"bottomright"}.}

\item{clusterId}{'character'.
Identification for the marker cluster layer.}
}
\value{
Used for the side-effect of a button placed on the web map.
}
\description{
These functions add buttons on a web map that control miscellaneous view options.
The \code{AddHomeButton} function sets the map view to a user specified geographical extent.
And the \code{AddClusterButton} function toggles marker clusters between frozen and unfrozen states.
}
\examples{
map <- CreateWebMap()
d <- maps::us.cities
opt <- leaflet::markerClusterOptions(showCoverageOnHover = FALSE)
id <- "cities_cluster"
map <- leaflet::addMarkers(map, lng = d$long, lat = d$lat, popup = d$name,
                           clusterOptions = opt, clusterId = id)
map <- AddHomeButton(map, extent = c(range(d$long), range(d$lat)))
map <- AddClusterButton(map, id)
map

}
\seealso{
\code{\link{CreateWebMap}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
