% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddWebMapButtons.R
\name{AddWebMapButtons}
\alias{AddWebMapButtons}
\alias{AddRefreshButton}
\alias{AddClusterButton}
\alias{AddSearchButton}
\title{Add Miscellaneous Web Map Buttons}
\usage{
AddRefreshButton(map, extent = NULL, position = "topleft")

AddClusterButton(map, clusterId, position = "topleft")

AddSearchButton(map, group, propertyName = "label", zoom = NULL,
  textPlaceholder = "Search...", position = "topleft")
}
\arguments{
\item{map}{'\link[leaflet]{leaflet}'.
Map widget object}

\item{extent}{'Spatial*', 'Raster*', 'Extent', 'matrix', or 'numeric'.
Extent object (or object from which an \code{\link[raster]{extent}} object can be extracted/created)
representing a rectangular geographical area on the map.
The extent must be specified in the coordinate reference system (CRS) of the web map,
usually in latitude and longitude using WGS 84 (also known as \href{https://epsg.io/4326}{EPSG:4326}).
By default, the extent object is read from the map widget.}

\item{position}{'character'.
Position of the button on the web map.
Possible values are \code{"topleft"}, \code{"topright"}, \code{"bottomleft"}, and \code{"bottomright"}.}

\item{clusterId}{'character'.
Identification for the marker cluster layer.}

\item{group}{'character'.
Name of the group whose features will be searched.}

\item{propertyName}{'character'.
Property name used to describe markers, such as, \code{"label"} and \code{"popup"}.}

\item{zoom}{'integer'.
Zoom level for move to location after marker found in search.}

\item{textPlaceholder}{'character'.
Text message to show in search element.}
}
\value{
Used for the side-effect of a button placed on a web map.
}
\description{
These functions add buttons on a \href{http://leafletjs.com/}{Leaflet} web map that control miscellaneous view options.
The \code{AddRefreshButton} function sets the map view to the original extent.
The \code{AddClusterButton} function toggles marker clusters between frozen and unfrozen states.
And the \code{AddSearchButton} function may be used to search for, and move to, the location of a marker.
}
\examples{
df <- maps::us.cities
spdf <- sp::SpatialPointsDataFrame(df[, c("long", "lat")], data = df,
                                   proj4string = sp::CRS("+init=epsg:4326"))
map <- CreateWebMap("Topo")
opt <- leaflet::markerClusterOptions(showCoverageOnHover = FALSE)
map <- leaflet::addMarkers(map, label = ~name, popup = ~name,
                           clusterOptions = opt, clusterId = "cluster",
                           group = "marker", data = spdf)
map <- AddRefreshButton(map)
map <- AddClusterButton(map, "cluster")
map <- AddSearchButton(map, "marker", zoom = 15, textPlaceholder = "Search cities...")
map

}
\seealso{
\code{\link{CreateWebMap}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
