% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutofitVariogram.R
\name{AutofitVariogram}
\alias{AutofitVariogram}
\title{Autofit Variogram Model}
\usage{
AutofitVariogram(formula, input_data, model = c("Sph", "Exp", "Gau", "Ste"),
  kappa = c(0.05, seq(0.2, 2, 0.1), 5, 10), fix.values = c(NA, NA, NA),
  verbose = FALSE, GLS.model = NA, start_vals = c(NA, NA, NA),
  bin_method = c("automap", "gstat", "equal_width", "equal_count"),
  merge.small.bins = TRUE, min.np.bin = 5, num.bins = NA,
  init.width = NA, ...)
}
\arguments{
\item{formula}{'formula'.
Formula that defines the dependent variable as a linear model of independent variables.}

\item{input_data}{'SpatialPointsDataFrame'.
Input data}

\item{model}{'character'.
Vector of model types that will be examined when automatically fitting a
theoretical variogram model to the empirical variogram data.
The different model types include:
\code{"Sph"}, spherical;
\code{"Exp"}, exponential;
\code{"Gau"}, Gaussian;
\code{"Ste"}, Matern with Michael Stein's parameterization.}

\item{kappa}{'numeric'.
Vector of smoothing parameter(s) of the Matern model.}

\item{fix.values}{'numeric'.
A vector of length 3 containing user specified fixed values for nugget, range, and sill.
Specifying value as \code{NA} indicates that the value is not fixed.}

\item{verbose}{'logical'.
Specify as true to print additional output describing the fitting process.}

\item{GLS.model}{'variogramModel'.
If specified, a Generalized Least Squares (GLS) empirical variogram is calculated.}

\item{start_vals}{'numeric'.
Starting values for variogram fitting, see \code{fix.values} argument for format.}

\item{bin_method}{'character'.
Binning method, either \code{"automap"} (default), \code{"gstat"}, \code{"equal_width"}, or \code{"equal_count"}.
See \sQuote{Details} section for method descriptions.}

\item{merge.small.bins}{'logical'.
Whether to check if there are bins with less than 5 pairs of observations.
If true, the first two bins are merged and the check is repeated.
This is repeated until all bins have more than \code{min.np.bin} observation pairs.}

\item{min.np.bin}{'integer'.
Minimum number of observation pairs that every bin should contain.}

\item{num.bins}{'integer'.
Initial number of bins}

\item{init.width}{'numeric'.
Initial bin width}

\item{...}{Parameters passed to the \code{\link[gstat]{variogram}} function when calculating the empirical variogram.}
}
\value{
Returns a 'autofitVariogram' object with the following components:
  \describe{
    \item{\code{exp_var}}{empirical variogram}
    \item{\code{var_model}}{theoretical variogram model}
    \item{\code{sserr}}{sums of squares between the empirical variogram and the fitted theoretical variogram model.}
  }
}
\description{
This function automatically fits a theoretical variogram model to input data.
Automatic fitting is performed using the \code{\link[gstat]{fit.variogram}} function.
And initial values for the sill, range, nugget, and model type are estimated using a
modified version of the \code{\link[automap]{autofitVariogram}} function.
Many options are available for defining lag distances (binning) in the empirical variogram.
}
\details{
The different binning methods include:
  \describe{
    \item{\code{"automap"}}{is the default binning method used by the \code{\link[automap]{autofitVariogram}} function.
      The distance intervals that define the bins are equal to 2, 4, 6, \ldots{}, 100 percent
      of about 1/3 the diagonal of the box spanning the observation locations.}
    \item{\code{"gstat"}}{is the default binning method used by the \code{\link[gstat]{variogram}} function.}
    \item{\code{"equal_width"}}{resizes the bins so that each contains the minimum number of observation pairs.}
    \item{\code{"equal_count"}}{defines bins so that each contains the same number of observation pairs.}
  }
}
\note{
Function source code was modified from the \code{autofitVariogram} function in the
  \href{https://CRAN.R-project.org/package=automap}{automap} package (version 1.0-14, license
  \href{https://cran.r-project.org/web/licenses/GPL-2}{GPL-2}) |
  \href{https://cran.r-project.org/web/licenses/GPL-3}{GPL-3}).
  And source code for the \code{"equal_width"} and \code{"equal_count"} binning methods was modified from the
  \href{http://hydroecology.net/know-your-variograms/}{know-your-variograms} blog post,
  by M.C. Koohafkan, accessed on Jan. 14, 2018.
}
\examples{
data(meuse, package = "sp")
sp::coordinates(meuse) <- ~x+y

v1 <- AutofitVariogram(zinc~1, meuse, verbose = TRUE)
automap:::plot.autofitVariogram(v1)

v2 <- AutofitVariogram(zinc~1, meuse, bin_method = "equal_width",
                       num.bins = 20, verbose = TRUE)
automap:::plot.autofitVariogram(v2)

v3 <- AutofitVariogram(zinc~1, meuse, bin_method = "equal_width",
                       init.width = 200, verbose = TRUE)
automap:::plot.autofitVariogram(v3)

v4 <- AutofitVariogram(zinc~1, meuse, bin_method = "equal_count",
                       min.np.bin = 800, verbose = TRUE)
automap:::plot.autofitVariogram(v4)

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{models}
