% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{inlmisc-deprecated}
\alias{inlmisc-deprecated}
\alias{GetTolColors}
\alias{PrintHelpPages}
\title{Deprecated Functions in the inlmisc Package}
\usage{
GetTolColors(...)

PrintHelpPages(...)
}
\arguments{
\item{...}{Arguments to be passed to the functions
(obscured to enforce the usage of new functions).}
}
\description{
Functions that are about to be removed from the \pkg{inlmisc} package.
}
\keyword{internal}
