% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid2Polygons.R
\name{FindPolyNodes}
\alias{FindPolyNodes}
\title{Find Polygon Nodes}
\usage{
FindPolyNodes(s)
}
\arguments{
\item{s}{'matrix'.
A 2-column table giving start- and end-node indexes for each segment in a level.}
}
\value{
Returns an object of class 'list'.
  Vector components giving node indexes for each polygon ring.
  The status of the polygon as a hole or an island is taken from the ring direction,
  with clockwise meaning island, and counter-clockwise meaning hole.
}
\description{
Find Polygon Nodes
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
