% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPval.R
\name{FormatPval}
\alias{FormatPval}
\title{Format P Values}
\usage{
FormatPval(x, digits = max(1, getOption("digits") - 2),
  eps = .Machine$double.eps, na.form = "NA", scientific = NA)
}
\arguments{
\item{x}{'numeric'.
Vector of p-values.}

\item{digits}{'integer'.
Number of signigicant digits to be used.}

\item{eps}{'numeric'.
Numerical tolerance}

\item{na.form}{'character'.
Value used for missing values.}

\item{scientific}{'logical'.
Indicates whether values should be encoded in scientific format.}
}
\description{
This function is intended for formatting p-values.
LaTeX formatting is used for scientific notation.
}
\examples{
x <- c(stats::runif(5), pi^-100, NA)
FormatPval(x)
format.pval(x)

x <- c(0.1, 0.0001, 1e-27)
FormatPval(x, scientific = TRUE)
FormatPval(x, digits = 3L, eps = 0.001)

}
\seealso{
\code{\link{ToScientific}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
