% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDaysInMonth.R
\name{GetDaysInMonth}
\alias{GetDaysInMonth}
\title{Get Number of Days in a Year and Month}
\usage{
GetDaysInMonth(x)
}
\arguments{
\item{x}{'integer' vector.
Year and month, with a required date format of \code{YYYYMM}.}
}
\value{
Returns a 'integer' vector indicating the number of days
  for each year and month value in \code{x}.
}
\description{
Calculate the number of days in a year and month.
}
\examples{
GetDaysInMonth(c("199802", "199804", "200412"))

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
