% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddInsetMap.R
\name{GetInsetLocation}
\alias{GetInsetLocation}
\title{Get Inset Location}
\usage{
GetInsetLocation(dx, dy, loc = "bottomleft", inset = 0.02, padx = 0,
  pady = 0)
}
\arguments{
\item{dx, dy}{'numeric'.
Width and height of the inset graphics, respectively.}

\item{loc}{'character'.
Single keyword used to specify the position of the inset in the main plot region:
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"},
or \code{"center"} to denote inset location.}

\item{inset}{'numeric'.
Inset distance(s) from the margins as a fraction of the main plot region.
Defaults to 2 percent of the axis range.}

\item{padx, pady}{'numeric'.
Padding distance in the \emph{x} and \emph{y} direction, respectively.}
}
\value{
Returns a 'numeric' vector of length 2 giving the user coordinates
  for the bottom-left corner of the inset.
}
\description{
This function determines the location for an inset in the main plot region.
}
\examples{
graphics::plot(NA, NA, xlim = c(0, 100), ylim = c(0, 100),
               xlab = "x", ylab = "y", xaxs = "i", yaxs = "i")
dx <- 20; dy <- 10
xy <- GetInsetLocation(dx, dy)
graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "red")
graphics::points(xy[1], xy[2], pch = 16)

xy <- GetInsetLocation(dx, dy, inset = 0.05)
graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "pink")
graphics::points(xy[1], xy[2], pch = 16)

xy <- GetInsetLocation(dx, dy, loc = "topright")
graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "blue")

xy <- GetInsetLocation(dx, dy, loc = "left", padx = 5)
graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "green")

xy <- GetInsetLocation(dx, dy, loc = "center")
graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "brown")

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
