% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetColors.R
\name{IsColor}
\alias{IsColor}
\title{Check whether a character string is a valid color specification.}
\usage{
IsColor(x, null.ok = FALSE)
}
\arguments{
\item{x}{'character' vector.
color specification}

\item{null.ok}{'logical' flag.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
A 'logical' vector of the same length as argument \code{x}.
}
\description{
Check whether a character string is a valid color specification.
}
\examples{
IsColor(c("red", "zzz", "#FFFFFF", "#7FAF1B111"))

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
