% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecreateLibrary.R
\name{IsPackageInstalled}
\alias{IsPackageInstalled}
\title{Check whether Package is Installed}
\usage{
IsPackageInstalled(x, lib = .libPaths())
}
\arguments{
\item{x}{'character'.
Vector of package names}

\item{lib}{'character'.
Vector of library trees}
}
\value{
A 'logical' vector the length of argument \code{x}.
}
\description{
This function checks whether a package(s) is installed under the library tree(s).
}
\examples{
IsPackageInstalled(c("inlmisc", "csimlni", "colorspace"))

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
