% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovePaletteMidpoint.R
\name{MovePaletteMidpoint}
\alias{MovePaletteMidpoint}
\title{Move Midpoint of Color Palette}
\usage{
MovePaletteMidpoint(x, midpoint = 0, buffer = 0, bias = TRUE,
  scheme = c("sunset", "BuRd", "PRGn"), alpha = NULL)
}
\arguments{
\item{x}{'numeric' object that can be passed to the \code{\link{range}} function
with \code{NA}'s removed.}

\item{midpoint}{'numeric' number.
Critical midpoint of data range, defaults to zero.}

\item{buffer}{'numeric' number.
Color level buffer around the critical midpoint measured as a fraction of the color range.}

\item{bias}{'logical' flag.
Whether to allow bias in the color spacing.}

\item{scheme}{'character' string.
Diverging color scheme name: specify \code{"sunset"}, \code{"BuRd"}, or \code{"PRGn"}.
Schemes are provided by the \code{\link{GetTolColors}} function.
Partial string matching is supported so argument may be abbreviated.}

\item{alpha}{'numeric' number.
Alpha transparency, values range from 0 (fully transparent) to 1 (fully opaque).
Specify as \code{NULL} to exclude the alpha channel value from colors.}
}
\value{
Returns a 'function' that takes an 'integer' argument (the required number of colors)
  and returns a 'character' vector of colors.
}
\description{
Move the critical midpoint of a diverging color palette.
}
\examples{
Fun <- MovePaletteMidpoint(c(-3, 7))
inlmisc:::plot.Tol(Fun(255))

# Range of a set of data (x)
n <- 20
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(x = c(-10,  0))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x = c( -7,  3))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x = c( -5,  5))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x = c( -3,  7))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x = c(  0, 10))(n))
par(op)

# Midpoint of data range (midpoint)
x <- c(-5, 5); n <- 255
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, midpoint = -6)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, midpoint = -2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, midpoint =  0)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, midpoint =  2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, midpoint =  6)(n))
par(op)

# Buffer around midpoint (buffer)
x <- c(-5, 5); n <- 20
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, buffer = 0.0)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, buffer = 0.2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, buffer = 0.4)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, buffer = 0.6)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, buffer = 0.8)(n))
par(op)

# Bias in color spacing (bias)
x <- c(-3, 7); n <- 20
op <- par(mfrow = c(2, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, bias = TRUE)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(x, bias = FALSE)(n))
par(op)

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{color}
