% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintHelpPages.R
\name{PrintHelpPages}
\alias{PrintHelpPages}
\title{Print Package Help Pages in HTML Format}
\usage{
PrintHelpPages(pkg, file = "", toc = FALSE, hr = TRUE,
  links = NULL)
}
\arguments{
\item{pkg}{'character' string.
Package name}

\item{file}{'character' string.
A connection, or a character string naming the file to append output to.
Prints to the standard output connection by default.}

\item{toc}{'logical' flag.
Whether to format the title of each help topic as a level 1 header.
The table of contents (toc) option in R Markdown requires Markdown headers.}

\item{hr}{'logical' flag.
Whether to add a horizontal rule or line to separate help pages.}

\item{links}{'character' vector.
Names of packages searched when creating internal hyperlinks to help topics
(an experimental feature).}
}
\description{
Print the HTML code associated with help pages of a loaded R package.
}
\examples{
\dontrun{
cat("---",
    "title: Help Topics",
    "output:",
    "  html_document:",
    "    toc: true",
    "    toc_float: true",
    "---\\n\\n---\\n",
    sep = "\\n", file = "help-example.Rmd")
PrintHelpPages("inlmisc", file = "help-example.Rmd", toc = TRUE)
rmarkdown::render("help-example.Rmd")
url <- file.path("file:/", getwd(), "help-example.html")
utils::browseURL(url)

file.remove("help-example.Rmd", "help-example.html")
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
