% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projargs}
\alias{projargs}
\title{Projection Arguments}
\format{A 'character' string with arguments specified as in the PROJ.4 documentation.}
\usage{
projargs
}
\description{
Projection arguments that define a custom coordinate reference system (CRS) sometimes used by the
U.S. Geological Survey Idaho National Laboratory Project Office.
Geographic coordinates are in units of meters, in conformance with the
North American Datum of 1927 (NAD 27), and placed in the
Albers Equal-Area Conic projection; standard parallels 42.83, 44.16; central meridian -113.00,
false easting 200,000 meters; false northing 0.
}
\examples{
crs <- sp::CRS(projargs)
print(crs)

# To build the dataset, run the following commands:

\dontrun{
projargs <- paste("+proj=aea",
                  "+lat_1=42.83333333333333 +lat_2=44.16666666666666",
                  "+lat_0=41.5 +lon_0=-113",
                  "+x_0=200000 +y_0=0",
                  "+ellps=clrk66",
                  "+datum=NAD27",
                  "+units=m",
                  "+no_defs")
save(projargs, file = "projargs.rda", compress = "xz")
}

}
\keyword{datasets}
