% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummariseBudget.R
\name{SummariseBudget}
\alias{SummariseBudget}
\title{Summarize MODFLOW Water Budget}
\usage{
SummariseBudget(budget, desc, id = NULL)
}
\arguments{
\item{budget}{character or list.
Either the path to a MODFLOW cell-by-cell budget file or
the value returned from the \code{\link{ReadModflowBinary}} function.}

\item{desc}{character.
Vector of data type descriptors, such as \code{c("wells", "drains")}.}

\item{id}{character.
Name of auxiliary variable, additional values associated with each cell.}
}
\value{
Returns a 'data.frame' object with the following variables:
  \describe{
    \item{desc}{data type description, such as "wells"}
    \item{kper}{stress period}
    \item{kstp}{time step}
    \item{id}{auxiliary variable name}
    \item{delt}{length of the current time step.}
    \item{pertim}{time in the stress period.}
    \item{totim}{total elapsed time}
    \item{count}{number of cells in each subset.}
    \item{flow.sum}{total volumetric flow rate}
    \item{flow.mean}{mean volumetric flow rate}
    \item{flow.median}{median volumetric flow rate}
    \item{flow.sd}{standard deviation of volumetric flow rate.}
    \item{flow.dir}{flow direction where "in" and "out" indicate
      water entering and leaving the groundwater system, respectively.}
  }
}
\description{
This is a utility function for \href{https://water.usgs.gov/ogw/modflow/}{MODFLOW}.
It summarizes volumetric flow rates by boundary condition types.
That is, it splits the MODFLOW water-budget data into subsets,
computes summary statistics for each, and returns the resulting summary table.
}
\details{
Subsets are grouped by the data type (\code{"desc"}), stress period
  (\code{"kper"}), time step (\code{"kstp"}), and optional auxiliary variable.
  The MODFLOW cell-by-cell budget file must be produced using the "COMPACT BUDGET" output option.
}
\examples{
\dontrun{
  bud <- SummariseBudget("modflow.bud", c("wells", "drains", "river leakage"), "id")
}

}
\seealso{
\code{\link{ReadModflowBinary}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
