% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_appendix.R
\name{usgs_appendix}
\alias{usgs_appendix}
\title{USGS Appendix Format}
\usage{
usgs_appendix(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a U.S. Geological Survey appendix.
}
\examples{
\dontrun{
rmarkdown::draft("Appendix.Rmd", template = "usgs_appendix", package = "inlmisc")
}

}
