% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_article.R
\name{usgs_article}
\alias{usgs_article}
\title{USGS Article Format}
\usage{
usgs_article(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a U.S. Geological Survey article.
}
\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "usgs_article", package = "inlmisc")
}

}
