% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIntervals.R
\name{AddIntervals}
\alias{AddIntervals}
\title{Add Vertical Intervals to Plot}
\usage{
AddIntervals(x, y0, y1, length = 0.03, code = 3,
  col = graphics::par("fg"), lty = graphics::par("lty"),
  lwd = graphics::par("lwd"), xpd = graphics::par("xpd"))
}
\arguments{
\item{x}{'numeric' or 'Date' vector.
\emph{x} coordinate of interval symbols.}

\item{y0}{'numeric' vector.
\emph{y} coordinate of points from which to draw.}

\item{y1}{'numeric' vector.
\emph{y} coordinate of points to which to draw.}

\item{length}{'numeric' number.
Horizontal length of the interval head, in inches.}

\item{code}{'integer' number.
Determines the type of interval to be drawn, see \sQuote{Details} section for code descriptions.}

\item{col, lty, lwd, xpd}{graphical parameters; see \code{\link{par}} for details.
\code{NA} values in \code{col} cause the interval to be omitted.}
}
\value{
Used for the side-effect of interval symbols drawn on the current graphics device.
}
\description{
Add vertical interval symbols (also known as error bars) to plots.
}
\details{
For each \code{i}, an interval is drawn between the
  point \code{(x[i], y0[i])} and the point \code{(x[i], y1[i])}.

  Code descriptions are as follows:
  \code{code = 0}, no interval head is drawn;
  \code{code = 1}, a head is drawn at \code{(x[i], y0[i])};
  \code{code = 2}, a head is drawn at \code{(x[i], y1[i])}; and
  \code{code = 3}, a head is drawn at both ends of the interval.
}
\examples{
set.seed(2)
x <- stats::runif(12)
y <- stats::rnorm(12)
plot(x, y)

dy <- sort(y) / 5
AddIntervals(x, y - dy, y + dy, col = "red", xpd = TRUE)

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
