% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddScaleBar.R
\name{AddScaleBar}
\alias{AddScaleBar}
\title{Add Scale Bar to Plot}
\usage{
AddScaleBar(asp = 1, unit = NULL, is.lonlat = FALSE,
  loc = c("bottomleft", "topleft", "topright", "bottomright"), offset = c(0,
  0), lab.vert.exag = NULL)
}
\arguments{
\item{asp}{'numeric'.
\emph{y/x} aspect ratio for spatial axes.}

\item{unit}{'character'.
Axis label for unit of measurement, for example "METERS".}

\item{is.lonlat}{'logical'.
Whether plot coordinates are in longitude and latitude.
If true, scale values are in units of kilometers.}

\item{loc}{'character'.
Position of the scale bar in the plot region:
\code{"bottomleft"}, \code{"topleft"}, \code{"topright"}, or
\code{"bottomright"} to denote scale location.}

\item{offset}{'numeric'.
The x and y adjustments of the scale bar, in inches.}

\item{lab.vert.exag}{'logical'.
Whether a label is drawn specifying the vertical exaggeration.}
}
\value{
Used for the side-effect of a scale bar drawn on the current graphics device.
}
\description{
This function can be used to add a scale bar to a plot.
}
\examples{
plot(-100:100, -100:100, type = "n", xlab = "x", ylab = "y", asp = 2)
AddScaleBar(2, unit = "FEET", loc = "topleft")
AddScaleBar(2, unit = "METERS", loc = "bottomright", offset = c(-0.2, 0))

}
\seealso{
\code{\link{PlotCrossSection}}, \code{\link{PlotMap}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
