% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddWebMapButtons.R
\name{AddWebMapButtons}
\alias{AddWebMapButtons}
\alias{AddRefreshButton}
\alias{AddClusterButton}
\alias{AddSearchButton}
\title{Add Miscellaneous Web Map Buttons}
\usage{
AddRefreshButton(map, extent = NULL, position = "topleft")

AddClusterButton(map, clusterId, position = "topleft")

AddSearchButton(map, group, propertyName = "label", zoom = 15,
  position = "topleft")
}
\arguments{
\item{map}{'\link[leaflet]{leaflet}'.
Map widget object}

\item{extent}{'Spatial*', 'Raster*', 'Extent', 'matrix', or 'numeric'.
Extent object (or object from which an \code{\link[raster]{extent}} object can be extracted/created)
representing a rectangular geographical area on the map.
The extent must be specified in the coordinate reference system (CRS) of the web map,
usually in latitude and longitude using WGS 84 (also known as \href{https://epsg.io/4326}{EPSG:4326}).
By default, the extent object is read from the map widget.}

\item{position}{'character'.
Position of the button on the web map.
Possible values are \code{"topleft"}, \code{"topright"}, \code{"bottomleft"}, and \code{"bottomright"}.}

\item{clusterId}{'character'.
Identification for the marker cluster layer.}

\item{group}{'character'.
Name of the group whose features will be searched.}

\item{propertyName}{'character'.
Property name used to describe markers, such as, \code{"label"} and \code{"popup"}.}

\item{zoom}{'integer'.
Zoom level for move to location after marker found in search.}
}
\value{
Used for the side-effect of a button placed on a web map.
}
\description{
These functions add buttons on a \href{http://leafletjs.com/}{Leaflet} web map that control miscellaneous view options.
The \code{AddRefreshButton} function sets the map view to the original extent.
The \code{AddClusterButton} function toggles marker clusters between frozen and unfrozen states.
And the \code{AddSearchButton} function may be used to search for, and move to, a marker.
}
\examples{
d <- maps::us.cities
map <- CreateWebMap("Topo")
opt <- leaflet::markerClusterOptions(showCoverageOnHover = FALSE)
map <- leaflet::addMarkers(map, d$long, d$lat, popup = d$name,
                           clusterOptions = opt, clusterId = "cities_cluster")
map <- AddRefreshButton(map)
map <- AddClusterButton(map, "cities_cluster")
map

map <- CreateWebMap("Imagery")
map <- leaflet::addMarkers(map, d$long, d$lat, popup = d$name, group = "marker")
map <- AddRefreshButton(map)
map <- AddSearchButton(map, "marker", propertyName = "popup")
map

}
\seealso{
\code{\link{CreateWebMap}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{hplot}
