% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildVignettes.R
\name{BuildVignettes}
\alias{BuildVignettes}
\title{Build Package Vignettes}
\usage{
BuildVignettes(pkg = ".", quiet = TRUE, gs_quality = NULL)
}
\arguments{
\item{pkg}{'character' string.
Package path, defaults to the working directory.}

\item{quiet}{'logical' flag.
Whether to supress most output.}

\item{gs_quality}{'character' string.
Quailty to use when compacting PDF files,
see \code{\link[tools]{compactPDF}} function for details.}
}
\description{
Build package vignettes from their sources.
}
\examples{
\dontrun{
BuildVignettes("<path/to/package>", gs_quality = "ebook")
}

}
\seealso{
\code{\link[tools]{buildVignettes}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
