% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildVignettes.R
\name{BuildVignettes}
\alias{BuildVignettes}
\title{Build Package Vignettes}
\usage{
BuildVignettes(pkg = ".", quiet = TRUE, clean = TRUE,
  gs_quality = NULL)
}
\arguments{
\item{pkg}{'character' string.
Package path, by default the \link[=getwd]{working directory}.}

\item{quiet}{'logical' flag.
Whether to suppress most output.}

\item{clean}{'logical' flag.
Whether to remove all intermediate files generated by the build.}

\item{gs_quality}{'character' string.
Quality of compacted PDF files,
see \code{\link[tools]{compactPDF}} function for details.}
}
\value{
Used for the side-effect of creating the PDF and (or) HTML versions
  of all vignettes and extracting the \R code.
}
\description{
Build package vignettes from their sources
(\code{.Rnw} and \code{.Rmd} files in the \code{vignettes} folder)
and place in the \code{inst/doc} folder.
}
\examples{
\dontrun{
BuildVignettes("<path/to/package>", gs_quality = "ebook")
}

}
\seealso{
\code{\link[tools]{buildVignettes}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
