% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildVignettes.R
\name{BuildVignettes}
\alias{BuildVignettes}
\title{Build Package Vignettes}
\usage{
BuildVignettes(dir = getwd(), doc = file.path(dir, "inst/doc"),
  gs_quality = c("none", "printer", "ebook", "screen"), clean = TRUE,
  quiet = TRUE)
}
\arguments{
\item{dir}{'character' string.
Path to a package's root source directory, by default the \link[=getwd]{working directory}.
The subdirectory \sQuote{vignettes} is searched for vignette source files.}

\item{doc}{'character' string.
Directory to write vignette and \emph{tangled} output files.}

\item{gs_quality}{'character' string.
Quality of compacted PDF files, the options are
\code{"none"} (no compression), \code{"printer"} (300 dpi),
\code{"ebook"} (150 dpi), and \code{"screen"} (72 dpi).}

\item{clean}{'logical' flag.
Whether to remove all intermediate files generated by the build.}

\item{quiet}{'logical' flag.
Whether to suppress most output.}
}
\value{
Used for the side-effect of creating the PDF and (or) HTML versions
  of all vignettes and extracting the \R code.
}
\description{
Build package vignettes from their source files.
}
\seealso{
\code{\link[tools]{buildVignettes}}, \code{\link[tools]{compactPDF}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
