% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOptimalSubset.R
\name{EncodeChromosome}
\alias{EncodeChromosome}
\alias{DecodeChromosome}
\title{Encode and Decode an Integer Chromosome}
\usage{
EncodeChromosome(x, res = 10L)

DecodeChromosome(y, res = 10L)
}
\arguments{
\item{x}{'numeric'.
Binary representation, a vector of \code{0}s and \code{1}s, of the integer chromosome.}

\item{res}{'integer'.
Resolution of integer chromosome in bits.}

\item{y}{'numeric'.
Integer chromosome, a vector of parameter values.}
}
\value{
\code{EncodeChromosome} returns a 'numeric' vector of \code{0}s and \code{1}s, that is,
    the binary representation of the integer chromosome.

  \code{DecodeChromosome} returns a 'numeric' vector of
    parameter values in the integer chromosome.
}
\description{
Functions for encoding and decoding an integer chromosome,
a set of integer values that define a proposed solution to the problem
that a genetic algorithm is trying to solve.
}
\examples{
string <- EncodeChromosome(c(4, 196, 67))
print(string)
print(DecodeChromosome(string))

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
