% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facilities}
\alias{facilities}
\title{Idaho National Laboratory Facilities}
\format{A 'SpatialPolygonsDataFrame' with 7 features and 1 variable.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1927 (NAD 27), and placed in the
  Albers Equal-Area Conic projection; standard parallels 42.83, 44.16; central meridian -113.00,
  false easting 200,000 meters; false northing 0.}
\source{
U.S. Geological Survey INL Project Office: \url{https://id.water.usgs.gov/INL/}
}
\usage{
facilities
}
\description{
Federal research facilities at the Idaho National Laboratory (\href{https://www.inl.gov/}{INL}).
}
\examples{
sp::plot(facilities, col = "black")
raster::text(facilities, facilities@data$NAME, cex = 0.6, pos = 1)
str(facilities@data)

# To build the dataset, run the following commands:

\dontrun{
f <- system.file("extdata", "facilities.zip", package = "inlmisc")
f <- utils::unzip(f, exdir = tempdir())
f <- grep(".shp$", f, value = TRUE)
layer <- tools::file_path_sans_ext(basename(f))
facilities <- rgdal::readOGR(dirname(f), layer, stringsAsFactors = FALSE)
facilities <- sp::spTransform(facilities, projargs)
save(facilities, file = "facilities.rda", compress = "xz")
}

}
\keyword{datasets}
