% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegionOfInterest.R
\name{GetRegionOfInterest}
\alias{GetRegionOfInterest}
\title{Get Region of Interest}
\usage{
GetRegionOfInterest(obj, alpha = NULL, width = 0, ...)
}
\arguments{
\item{obj}{'SpatialPoints*'.
Spatial points}

\item{alpha}{'numeric' number.
Value of alpha, used to implement a generalization of the convex hull
(Edelsbrunner and others, 1983).
Requires that the \pkg{alphahull} and \pkg{maptools} packages are available.}

\item{width}{'numeric' number.
Buffer distance from geometry of convex hull.}

\item{...}{Additional arguments to be passed to the \code{\link[rgeos]{gBuffer}} function.}
}
\value{
Returns an object of class 'SpatialPolygons'.
}
\description{
Create a spatial polygon describing the convex hull of a set of spatial points.
}
\examples{
set.seed(123)
pts <- sp::SpatialPoints(cbind(x = stats::runif(50), y = stats::runif(50)))
sp::plot(GetRegionOfInterest(pts, width = 0.05), border = "red", lty = 2)
sp::plot(GetRegionOfInterest(pts), border = "red", add = TRUE)
sp::plot(GetRegionOfInterest(pts, alpha = 0.5), border = "blue", add = TRUE)
sp::plot(pts, add = TRUE)

n <- 300
set.seed(321)
theta <- stats::runif(n, 0, 2 * pi)
r <- sqrt(stats::runif(n, 0.25^2, 0.50^2))
x <- cbind(0.5 + r * cos(theta), 0.5 + r * sin(theta))
pts <- sp::SpatialPoints(x)
sp::plot(GetRegionOfInterest(pts, alpha = 0.1, width = 0.05), col = "green")
sp::plot(GetRegionOfInterest(pts, alpha = 0.1), col = "yellow", add = TRUE)
sp::plot(pts, add = TRUE)

}
\references{
Edelsbrunner, H., Kirkpatrick, D.G. and Seidel, R., 1983,
  On the shape of a set of points in the plane:
  IEEE Transactions on Information Theory, v. 29, no. 4, p. 551--559.
}
\seealso{
\code{\link[grDevices]{chull}}, \code{\link[alphahull]{ashape}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
