% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegionOfInterest.R
\name{GetRegionOfInterest}
\alias{GetRegionOfInterest}
\title{Get Region of Interest}
\usage{
GetRegionOfInterest(obj, width = 0, ...)
}
\arguments{
\item{obj}{'SpatialPoints*'.
Spatial points}

\item{width}{'numeric' number.
Buffer distance from geometry of convex hull.}

\item{...}{Additional arguments to be passed to the \code{\link[rgeos]{gBuffer}} function.}
}
\value{
Returns an object of class 'SpatialPolygons'.
}
\description{
Create a spatial polygon describing the convex hull of a set of spatial points.
}
\examples{
set.seed(123)
pts <- sp::SpatialPoints(cbind(x = stats::runif(50), y = stats::runif(50)))
sp::plot(GetRegionOfInterest(pts, width = 0.05), lty = 2)
sp::plot(GetRegionOfInterest(pts), add = TRUE)
sp::plot(pts, col = "red", add = TRUE)

}
\seealso{
\code{\link[grDevices]{chull}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
