% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegionOfInterest.R
\name{GetRegionOfInterest}
\alias{GetRegionOfInterest}
\title{Get Region of Interest}
\usage{
GetRegionOfInterest(obj, alpha = NULL, width = 0, ...)
}
\arguments{
\item{obj}{'SpatialPoints*'.
Sample of points}

\item{alpha}{'numeric' number.
Value of \eqn{\alpha}, used to implement a generalization of the convex hull
(Edelsbrunner and others, 1983).
As \eqn{\alpha} decreases, the shape shrinks.
Requires that the \pkg{alphahull} and \pkg{maptools} packages are available.
Note that the \href{https://CRAN.R-project.org/package=alphahull}{alphahull} package
is released under a restrictive non-free software license.}

\item{width}{'numeric' number.
Buffer distance from geometry of convex hull.}

\item{...}{Additional arguments to be passed to the \code{\link{gBuffer}} function.}
}
\value{
Returns an object of class 'SpatialPolygons'.
}
\description{
Create a spatial polygon describing the convex hull of a set of spatial points.
}
\examples{
set.seed(123)

n <- 50
pts <- sp::SpatialPoints(cbind(x = stats::runif(n), y = stats::runif(n)))
sp::plot(GetRegionOfInterest(pts, width = 0.05), border = "blue", lty = 2)
sp::plot(GetRegionOfInterest(pts), border = "red", add = TRUE)
sp::plot(GetRegionOfInterest(pts, width = -0.05), lty = 2, add = TRUE)
sp::plot(pts, add = TRUE)

\dontrun{
n <- 300
theta <- stats::runif(n, 0, 2 * pi)
r <- sqrt(stats::runif(n, 0.25^2, 0.50^2))
pts <- sp::SpatialPoints(cbind(0.5 + r * cos(theta), 0.5 + r * sin(theta)))
sp::plot(GetRegionOfInterest(pts, alpha = 0.1, width = 0.05), col = "green")
sp::plot(GetRegionOfInterest(pts, alpha = 0.1), col = "yellow", add = TRUE)
sp::plot(pts, add = TRUE)
}

}
\references{
Edelsbrunner, H., Kirkpatrick, D.G. and Seidel, R., 1983,
  On the shape of a set of points in the plane:
  IEEE Transactions on Information Theory, v. 29, no. 4, p. 551--559.
}
\seealso{
\code{\link[grDevices]{chull}}, \code{\link[alphahull]{ashape}},
  \code{\link[maptools]{checkPolygonsHoles}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
