% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTolColors.R
\name{GetTolColors}
\alias{GetTolColors}
\title{Tol Color Schemes}
\usage{
GetTolColors(n, scheme = "smooth rainbow", alpha = NULL, start = 0,
  end = 1, bias = 1, reverse = FALSE, plot = FALSE)
}
\arguments{
\item{n}{'integer'.
Number of colors to be in the palette, the maximum value is
dependent on the specified color scheme, see \sQuote{Details} section.}

\item{scheme}{'character'.
Color scheme: select
  \code{"bright"}, \code{"vibrant"}, \code{"muted"}, \code{"pale"},
  \code{"dark"}, or \code{"light"} for qualitative colors;
  \code{"sunset"}, \code{"BuRd"}, or \code{"PRGn"} for diverging colors;
  \code{"YlOrBr"}, \code{"discrete rainbow"} or \code{"smooth rainbow"} for sequential colors; and
  \code{"ground cover"} for the \href{http://glcf.umd.edu/data/landcover/data.shtml}{AVHRR}
  global land cover classification (Hansen and others, 1998).}

\item{alpha}{'numeric'.
Alpha transparency, values range from 0 (fully transparent) to 1 (fully opaque).
Specify as \code{NULL} to exclude the alpha channel value from the color name.}

\item{start, end}{'numeric'.
Starting and ending color level in the palette, respectively.
Specified as a number in the interval from 0 to 1.
Applies only to interpolated color schemes:
\code{"sunset"}, \code{"BuRd"}, \code{"PRGn"}, \code{"YlOrBr"}, and \code{"smooth rainbow"}.}

\item{bias}{'numeric'.
Interpolation bias where larger values result in more widely spaced colors at the high end.
See \code{\link[grDevices]{colorRamp}} function for details.}

\item{reverse}{'logical'.
Whether to reverse the direction of colors in the palette.}

\item{plot}{'logical'.
Whether to display the color palette.}
}
\value{
Returns a 'character' vector of \code{n} color names---each
  name is 7 or 9 characters and formatted as
  \code{"#"} followed by the red, blue, green, and optionally alpha values in hexadecimal.
  For some color schemes the returned object includes a \code{"bad"} attribute giving
  the color assigned to bad data.
}
\description{
This function creates a vector of \code{n} colors from
qualitative, diverging, and sequential color schemes by Paul Tol (2018).
}
\details{
Limits on the maximum number of discrete colors for a scheme are:
  \code{n < 8} for \code{"bright"} and \code{"vibrant"};
  \code{n < 10} for \code{"muted"}, \code{"light"}, \code{"YlOrBr"}, \code{"BuRd"}, and \code{"PRGn"};
  \code{n < 24} for \code{"discrete rainbow"}.
  Schemes \code{"pale"} (\code{n = 6}),  \code{"dark"} (\code{n = 6}),
  and \code{"ground cover"} (\code{n = 14}) are intended to be
  accessed in their entirety and subset using color names.
}
\examples{

col <- GetTolColors(3); print(col)
attr(col, "bad")

# Number of colors (n)
op <- par(mfrow = c(7, 1), oma = c(0, 0, 0, 0))
GetTolColors( 2, scheme = "discrete rainbow", plot = TRUE)
GetTolColors( 4, scheme = "discrete rainbow", plot = TRUE)
GetTolColors( 6, scheme = "discrete rainbow", plot = TRUE)
GetTolColors( 8, scheme = "discrete rainbow", plot = TRUE)
GetTolColors(10, scheme = "discrete rainbow", plot = TRUE)
GetTolColors(15, scheme = "discrete rainbow", plot = TRUE)
GetTolColors(23, scheme = "discrete rainbow", plot = TRUE)
par(op)

# Qualitative color schemes (scheme)
op <- par(mfrow = c(6, 1), oma = c(0, 0, 0, 0))
GetTolColors(7, scheme = "bright",  plot = TRUE)
GetTolColors(7, scheme = "vibrant", plot = TRUE)
GetTolColors(9, scheme = "muted",   plot = TRUE)
GetTolColors(6, scheme = "pale",    plot = TRUE)
GetTolColors(6, scheme = "dark",    plot = TRUE)
GetTolColors(9, scheme = "light",   plot = TRUE)
par(op)

# Diverging color schemes (scheme)
op <- par(mfrow = c(6, 1), oma = c(0, 0, 0, 0))
GetTolColors( 11, scheme = "sunset", plot = TRUE)
GetTolColors(255, scheme = "sunset", plot = TRUE)
GetTolColors(  9, scheme = "BuRd",   plot = TRUE)
GetTolColors(255, scheme = "BuRd",   plot = TRUE)
GetTolColors(  9, scheme = "PRGn",   plot = TRUE)
GetTolColors(255, scheme = "PRGn",   plot = TRUE)
par(op)

# Sequential color schemes (scheme)
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
GetTolColors(  9, scheme = "YlOrBr",           plot = TRUE)
GetTolColors(255, scheme = "YlOrBr",           plot = TRUE)
GetTolColors( 23, scheme = "discrete rainbow", plot = TRUE)
GetTolColors( 34, scheme = "smooth rainbow",   plot = TRUE)
GetTolColors(255, scheme = "smooth rainbow",   plot = TRUE)
par(op)

# Cover color scheme (scheme)
op <- par(oma = c(1, 0, 0, 0), cex = 0.7)
GetTolColors(14, scheme = "ground cover", plot = TRUE)
par(op)

# Alpha transparency (alpha)
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
GetTolColors(34, alpha = 1.0, plot = TRUE)
GetTolColors(34, alpha = 0.8, plot = TRUE)
GetTolColors(34, alpha = 0.6, plot = TRUE)
GetTolColors(34, alpha = 0.4, plot = TRUE)
GetTolColors(34, alpha = 0.2, plot = TRUE)
par(op)

# Color levels (start, end)
op <- par(mfrow = c(4, 1), oma = c(0, 0, 0, 0))
GetTolColors(255, start = 0.0, end = 1.0, plot = TRUE)
GetTolColors(255, start = 0.0, end = 0.5, plot = TRUE)
GetTolColors(255, start = 0.5, end = 1.0, plot = TRUE)
GetTolColors(255, start = 0.3, end = 0.9, plot = TRUE)
par(op)

# Interpolation bias (bias)
op <- par(mfrow = c(7, 1), oma = c(0, 0, 0, 0))
GetTolColors(255, bias = 0.4, plot = TRUE)
GetTolColors(255, bias = 0.6, plot = TRUE)
GetTolColors(255, bias = 0.8, plot = TRUE)
GetTolColors(255, bias = 1.0, plot = TRUE)
GetTolColors(255, bias = 1.2, plot = TRUE)
GetTolColors(255, bias = 1.4, plot = TRUE)
GetTolColors(255, bias = 1.6, plot = TRUE)
par(op)

# Reverse colors (reverse)
op <- par(mfrow = c(2, 1), oma = c(0, 0, 0, 0), cex = 0.7)
GetTolColors(10, reverse = FALSE, plot = TRUE)
GetTolColors(10, reverse = TRUE,  plot = TRUE)
par(op)

}
\references{
Hansen, M., DeFries, R., Townshend, J.R.G., and Sohlberg, R., 1998,
  UMD Global Land Cover Classification, 1 Kilometer, 1.0:
  Department of Geography, University of Maryland, College Park, Maryland, 1981-1994.

  Tol, Paul, 2018, Colour Schemes:
  SRON Technical Note, doc. no. SRON/EPS/TN/09-002, issue 3.0, 17 p.,
  accessed July 18, 2018 at \url{https://personal.sron.nl/~pault/data/colourschemes.pdf}.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{color}
