% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inl}
\alias{inl}
\title{Idaho National Laboratory Boundary}
\format{A 'SpatialPolygonsDataFrame' object with 1 feature and 4 variables.
  See \code{\link{projargs}} dataset for coordinate reference system.}
\source{
U.S. Geological Survey INL Project Office: \url{https://id.water.usgs.gov/INL/}
}
\usage{
inl
}
\description{
The political boundary of the Idaho National Laboratory (\href{https://www.inl.gov/}{INL}).
}
\examples{
sp::plot(inl)
str(inl@data)

# To build the dataset, run the following commands:

\dontrun{
f <- system.file("extdata", "inl.zip", package = "inlmisc")
f <- utils::unzip(f, exdir = tempdir())
f <- grep(".shp$", f, value = TRUE)
inl <- rgdal::readOGR(dirname(f), tools::file_path_sans_ext(basename(f)))
inl <- sp::spTransform(inl, projargs)
save(inl, file = "inl.rda", compress = "xz")
}

}
\keyword{datasets}
