% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecreateLibrary.R
\name{IsPackageInstalled}
\alias{IsPackageInstalled}
\title{Check whether Package is Installed}
\usage{
IsPackageInstalled(x, lib = .libPaths())
}
\arguments{
\item{x}{'character' vector.
Package names}

\item{lib}{'character' vector.
Library trees}
}
\value{
A 'logical' vector of the same length as argument \code{x}.
}
\description{
Check whether a package(s) is installed under the library tree(s).
}
\examples{
IsPackageInstalled(c("inlmisc", "csimlni", "devtools"))

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
