% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovePaletteMidpoint.R
\name{MovePaletteMidpoint}
\alias{MovePaletteMidpoint}
\title{Move Midpoint of Color Palette}
\usage{
MovePaletteMidpoint(ran, mid = 0, scheme = c("sunset", "BuRd", "PRGn"),
  buffer = 0, bias = TRUE)
}
\arguments{
\item{ran}{'numeric' vector of length 2 or more.
Data range or values that will be used to calculate the range.}

\item{mid}{'numeric' number.
Critical midpoint of data range, defaults to 0.}

\item{scheme}{'character' string.
Diverging color scheme: specify \code{"sunset"}, \code{"BuRd"}, or \code{"PRGn"}.
Partial string matching is supported so argument may be abbreviated.}

\item{buffer}{'numeric' number.
Color level buffer around the critical midpoint measured as a fraction of the data range.}

\item{bias}{'logical' flag.
Whether to allow bias in the color spacing.}
}
\value{
Returns a 'function' that takes an 'integer' argument (the required number of colors)
  and returns a 'character' vector of colors.
}
\description{
Move the critical midpoint of a diverging color palette.
}
\examples{
Fun <- MovePaletteMidpoint(ran = c(-3, 7))
inlmisc:::plot.Tol(Fun(255))

# Data range (ran)
n <- 10
op <- par(mfrow = c(3, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran = c( -5,  5))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran = c(-10,  0))(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran = c(  0, 10))(n))
par(op)

# Color schemes (scheme)
ran <- c(-5, 5); n <- 255
op <- par(mfrow = c(3, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, scheme = "sunset")(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, scheme = "BuRd")(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, scheme = "PRGn")(n))
par(op)

# Midpoint of data range (mid)
ran <- c(-5, 5); n <- 255
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, mid = -6)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, mid = -2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, mid =  0)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, mid =  2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, mid =  6)(n))
par(op)

# Buffer around midpoint (buffer)
ran <- c(-5, 5); n <- 10
op <- par(mfrow = c(5, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, buffer = 0.0)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, buffer = 0.2)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, buffer = 0.4)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, buffer = 0.6)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, buffer = 0.8)(n))
par(op)

# Bias in color spacing (bias)
ran <- c(-3, 7); n <- 20
op <- par(mfrow = c(2, 1), oma = c(0, 0, 0, 0))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, bias = TRUE)(n))
inlmisc:::plot.Tol(MovePaletteMidpoint(ran, bias = FALSE)(n))
par(op)

}
\seealso{
\code{\link{GetTolColors}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{color}
