% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintHelpPages.R
\name{PrintHelpPages}
\alias{PrintHelpPages}
\title{Print Package Help Pages in HTML Format}
\usage{
PrintHelpPages(pkg, file = "", toc = FALSE, hr = TRUE,
  links = NULL)
}
\arguments{
\item{pkg}{'character' string.
Package name}

\item{file}{'character' string.
A connection, or a character string naming the file to print to.
Prints to the standard output connection by default.}

\item{toc}{'logical' flag.
Whether to format the initial HTML header of each help page in Markdown.
The table of contents (toc) option in R Markdown requires Markdown headers.}

\item{hr}{'logical' flag.
Whether to add a horizontal rule or line to separate help pages.}

\item{links}{'character' vector (experimental).
Names of packages searched when creating internal hyperlinks to help topics.}
}
\description{
Print the HTML code associated with help pages of a loaded R package.
}
\examples{
x <- utils::capture.output(PrintHelpPages("inlmisc", toc = TRUE))
cat(x, file = "help-example.Rmd", sep = "\\n")
rmarkdown::render("help-example.Rmd")
\donttest{utils::browseURL(sprintf("file://\%s", file.path(getwd(), "help-example.html")))}

file.remove("help-example.Rmd", "help-example.html")

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
