% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintHelpPages.R
\name{PrintHelpPages}
\alias{PrintHelpPages}
\title{Print Package Help Pages in HTML Format}
\usage{
PrintHelpPages(pkg, toc = FALSE, hr = TRUE)
}
\arguments{
\item{pkg}{'character' string.
Package name}

\item{toc}{'logical' flag.
Whether to format the initial HTML header of each help page in Markdown.
The table of contents (toc) option in R Markdown requires Markdown headers.}

\item{hr}{'logical' flag.
Whether to add a horizontal rule (hr) or line to separate help pages.}
}
\description{
Print the HTML code associated with help pages of a loaded R package.
}
\examples{
PrintHelpPages("inlmisc")

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
