% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintHelpPages.R
\name{PrintHelpPages}
\alias{PrintHelpPages}
\title{Print Package Help Pages in HTML Format}
\usage{
PrintHelpPages(pkg, file = "", internal = FALSE, sep = "<hr>",
  links = NULL, ...)
}
\arguments{
\item{pkg}{'character' vector.
Package name(s)}

\item{file}{'connection' or 'character' string.
Names the file to append output to.
Prints to the standard output connection by default.}

\item{internal}{'logical' flag.
Whether to print help topics flagged with the keyword "internal".}

\item{sep}{'character' string.
HTML to separate help topics, a horizontal line by default.}

\item{links}{'character' vector (experimental).
Names of packages searched when creating internal hyperlinks to help topics.}

\item{...}{Not used}
}
\value{
Invisible \code{NULL}
}
\description{
Print the HTML code associated with help pages of one or more add-on packages.
}
\examples{
\dontrun{
cat("---",
    "output:",
    "  html_document:",
    "    toc: true",
    "    toc_float:",
    "      smooth_scroll: false",
    "    mathjax: null",
    "---",
    sep = "\\n", file = "help-example.Rmd")
PrintHelpPages("inlmisc", file = "help-example.Rmd")
rmarkdown::render("help-example.Rmd")
url <- file.path("file:/", getwd(), "help-example.html")
utils::browseURL(url)

file.remove("help-example.Rmd", "help-example.html")
}

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
