% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTable.R
\name{PrintTable}
\alias{PrintTable}
\title{Print as LaTeX Table}
\usage{
PrintTable(d, colheadings = NULL, align = NULL, label = NULL,
  title = NULL, headnotes = NULL, footnotes = NULL, nrec = nrow(d),
  na = "--", rm_dup = NULL)
}
\arguments{
\item{d}{'data.frame'.
Something.}

\item{colheadings}{'character'.
Something.}

\item{align}{'character'.
Something.}

\item{label}{'character'.
Something.}

\item{title}{'character'.
Something.}

\item{headnotes}{'character'.
Something.}

\item{footnotes}{'character'.
Something.}

\item{nrec}{'integer'.
Something.}

\item{na}{'character'.
Something.}

\item{rm_dup}{'character' or 'integer'.
Something.}
}
\description{
This function prints the LaTeX code associated with the supplied data frame.
}
\details{
Requires \code{\\usepackage{caption}}, \code{\\usepackage{booktabs}},
  and \code{\\usepackage{makecell}} in the LaTeX preamble.
}
\examples{
d <- datasets::iris
d <- d[, c("Species", "Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
colheadings <- c("Species of \\\\\\\\ Iris",
                 "Sepal \\\\\\\\ length \\\\\\\\ (cm)", "Sepal \\\\\\\\ width \\\\\\\\ (cm)",
                 "Petal \\\\\\\\ length \\\\\\\\ (cm)", "Petal \\\\\\\\ width \\\\\\\\ (cm)")
align <- c("l", "r", "r", "r", "r")
title <- "Measurments of sepal length and width and petal length and width,
          for three species of Iris flower."
headnotes <- "\\\\textbf{Species of Iris}: inlcudes setosa, versicolor, and virginica.
              \\\\textbf{Abbreviations}: cm, centimeters"
PrintTable(d, colheadings, align, title = title, headnotes = headnotes,
           nrec = c(40L, 45L), rm_dup = 1L)

}
\seealso{
\code{\link[xtable]{xtable}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{print}
