% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTable.R
\name{PrintTable}
\alias{PrintTable}
\title{Print as LaTeX Table}
\usage{
PrintTable(
  d,
  colheadings = NULL,
  align = NULL,
  digits = NULL,
  label = NULL,
  title = NULL,
  headnotes = NULL,
  footnotes = NULL,
  nrec = nrow(d),
  hline = NULL,
  na = "\\\\textemdash",
  rm_dup = NULL,
  landscape = FALSE,
  ...
)
}
\arguments{
\item{d}{'data.frame' or 'matrix'.
Data table to print.}

\item{colheadings}{'character' vector, 'matrix', or 'data.frame'.
Column headings.
For table objects, rows represent layers of headings (stacked headings).
The number of columns (or vector length) must equal the number of columns in argument \code{d}.
A column heading can span multiple columns by repeating adjacent headings.
Use \code{\\\\\\\\} to code a line break.}

\item{align}{'character' vector.
Column alignment.
Specify \code{"l"} to left align, \code{"r"} to right align, \code{"c"} to center align,
and \code{"S"} to align on the decimal point.}

\item{digits}{'integer' vector.
Number of digits to display in the corresponding columns.}

\item{label}{'character' string.
LaTeX label anchor.
Specifying this argument allows you to easily reference the table within the LaTeX document.
For example, when \code{label = "id"}, use \code{\\ref\{id\}}
to reference the table within a sentence.}

\item{title}{'character' string.
Table caption}

\item{headnotes}{'character' string.
Label placed below the table caption to provide information pertaining to the caption,
to the table as a whole, or to the column headings.}

\item{footnotes}{'character' string.
Label placed at the end of the table to provide explanations
of individual entries in the table.}

\item{nrec}{'integer' vector of length 1 or 2, value is recycled as necessary.
Maximum number of records to show on the first page, and every subsequent page, respectively.}

\item{hline}{'integer' vector.
Numbers between 1 and \code{nrow(d) - 1} indicating the table rows after which
a horizontal line should appear.}

\item{na}{'character' string.
Value to be used for missing values in table entries.}

\item{rm_dup}{'integer' count.
End value of a sequence of column indexes \code{(1:rm_dup)}.
Duplicate values contained in these columns will be set equal to an empty string.
Where duplicates in a column are determined from the 'character' vector formed by
combining its content with the content from all previous columns in the table.}

\item{landscape}{'logical' flag.
If true, conforming PDF viewers will display the table in landscape orientation.
This option requires \code{\\usepackage[pdftex]{lscape}} in the LaTeX preamble.}

\item{...}{Additional arguments to be passed to the \code{\link[xtable]{print.xtable}} function.
The arguments \code{type}, \code{hline.after} and \code{add.to.row} should not be included.}
}
\value{
Invisible \code{NULL}
}
\description{
Print the LaTeX code associated with the supplied data table.
The applied output format attempts to adhere to the design recommendations
for tables in United States Geological Survey (USGS) publications.
}
\details{
Requires \code{\\usepackage{caption}}, \code{\\usepackage{booktabs}},
  \code{\\usepackage{makecell}}, \code{\\usepackage{multirow}}, and
  \code{\\usepackage{siunitx}} in the LaTeX preamble.
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{print}
