% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTable.R
\name{PrintTable}
\alias{PrintTable}
\title{Print as LaTeX Table}
\usage{
PrintTable(d, colheadings = NULL, align = NULL, digits = NULL,
  label = NULL, title = NULL, headnotes = NULL, footnotes = NULL,
  nrec = nrow(d), hline = NULL, na = "--", rm_dup = NULL,
  landscape = FALSE)
}
\arguments{
\item{d}{'data.frame'.
Data table to print.}

\item{colheadings}{'character'.
Vector of length equal to the number of columns in the table, specifying column headers.}

\item{align}{'character'.
Vector of length equal to the number of columns in the table,
indicating the alignment of the corresponding columns.
Use \code{"l"}, \code{"r"}, and \code{"c"} to denote left, right, and center alignment, respectively.}

\item{digits}{'integer'.
Vector of length equal to the number of columns in the table,
indicating the number of digits to display in the corresponding columns.}

\item{label}{'character'.
String containing the LaTeX label anchor.}

\item{title}{'character'.
String containing the table caption.}

\item{headnotes}{'character'.
String placed below the \code{title} to provide information pertaining to the title,
to the table as a whole, or to the column headings.}

\item{footnotes}{'character'.
String placed at the end of the table to provide explanations of individual entries in the table.}

\item{nrec}{'integer'.
End value of a sequence of column indexes with a starting value of \code{1},
specifying columns to remove duplicate values.}

\item{hline}{'integer'.
Vector of numbers between 1 and \code{nrow(d) - 1}, indicating the rows after which
a horizontal line should appear.}

\item{na}{'character'.
String to be used for missing values in table entries.}

\item{rm_dup}{'integer'.
Vector of length equal to 2, indicating the maximum number of records to show on the first page,
and every subsequent page, respectively.
Value is recycled as necessary.}

\item{landscape}{'logical'.
If true, displays the table in landscape orientation by conforming PDF viewers.
This option requires \code{\\usepackage[pdftex]{lscape}} in the LaTeX preamble.}
}
\description{
This function prints the LaTeX code associated with the supplied data table.
}
\details{
Requires \code{\\usepackage{caption}}, \code{\\usepackage{booktabs}}, and
  \code{\\usepackage{makecell}} in the LaTeX preamble.
}
\examples{
d <- datasets::iris
d <- d[, c("Species", "Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")]
colheadings <- c("Species of \\\\\\\\ Iris",
                 "Sepal \\\\\\\\ length \\\\\\\\ (cm)", "Sepal \\\\\\\\ width \\\\\\\\ (cm)",
                 "Petal \\\\\\\\ length \\\\\\\\ (cm)", "Petal \\\\\\\\ width \\\\\\\\ (cm)")
align <- c("l", "r", "r", "r", "r")
title <- "Measurments of sepal length and width and petal length and width,
          for three species of Iris flower."
headnotes <- "\\\\textbf{Species of Iris}: inlcudes setosa, versicolor, and virginica.
              \\\\textbf{Abbreviations}: cm, centimeters"
PrintTable(d, colheadings, align, title = title, headnotes = headnotes,
           nrec = c(40L, 45L), rm_dup = 1L)

}
\seealso{
\code{\link[xtable]{xtable}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{print}
