% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projargs}
\alias{projargs}
\title{Projection Arguments}
\format{A 'character' string with projection arguments in PROJ.4 format.}
\usage{
projargs
}
\description{
Projection arguments that define a custom coordinate reference system (CRS) used by the
U.S. Geological Survey Idaho National Laboratory Project Office.
The CRS consists of all the following components:
Albers Equal-Area Conic projection;
latitude of first and second standard parallels is 42.83 and 44.16, respectively;
latitude of origin is 41.5;
central meridian is -113;
false easting and northing is 200,000 and 0 meters, respectively;
Clarke (1966) reference ellipsoid;
North American Datum of 1927; and
units of meters.
}
\examples{
crs <- sp::CRS(projargs)
print(crs)

# To build the dataset, run the following commands:

\dontrun{
projargs <- paste("+proj=aea",
                  "+lat_1=42.83333333333333 +lat_2=44.16666666666666",
                  "+lat_0=41.5 +lon_0=-113",
                  "+x_0=200000 +y_0=0",
                  "+ellps=clrk66",
                  "+datum=NAD27",
                  "+units=m",
                  "+no_defs")
save(projargs, file = "projargs.rda", compress = "xz")
}

}
\keyword{datasets}
