% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModflowBinary.R
\name{ReadModflowBinary}
\alias{ReadModflowBinary}
\title{Read MODFLOW Binary File}
\usage{
ReadModflowBinary(path, data.type = c("array", "flow"), rm.totim.0 = FALSE)
}
\arguments{
\item{path}{character.
Path name of the binary file.}

\item{data.type}{character.
Description of how the data were saved.
Specify \code{"array"} for array data (such as heads) and
\code{"flow"} for flow data (such as cell-by-cell flows).}

\item{rm.totim.0}{logical.
If true, stress-periods at time zero are removed.}
}
\value{
Returns a \code{list} object of length equal to the
  number of times the data type is written to the binary file.
  The following list components are returned:
  \describe{
    \item{d}{matrix of data values.
      The format typically coincides with the horizontal model grid.
      The one exception is when \code{data.type = "flow"} and the
      cell-by-cell budget file was saved using the "COMPACT BUDGET" output option.
      The compact format stores values associated with each cell:
      columns include cell index (\code{"icell"}), model layer (\code{"layer"}),
      model row (\code{"row"}), model column (\code{"column"}), cell-by-cell flow (\code{"flow"}),
      and any auxiliary variables saved using the "AUXILIARY" output option.}
    \item{kstp}{time step}
    \item{kper}{stress period}
    \item{desc}{data-type description}
    \item{ilay}{model-grid layer}
    \item{delt}{length of the current time step.}
    \item{pertim}{time in the stress period.}
    \item{totim}{total elapsed time}
  }
}
\description{
This is a utility function for \href{https://water.usgs.gov/ogw/modflow/}{MODFLOW}.
It reads data from binary files produced by MODFLOW.
}
\details{
This function reads head, drawdown,
  and cell budget files generated by MODFLOW.
}
\examples{
path <- system.file("extdata", "ex3D.hds", package = "inlmisc")
hds <- ReadModflowBinary(path, "array")
image(hds[[1]]$d)

}
\seealso{
\code{\link{SummariseBudget}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
