% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModflowBinary.R
\name{ReadModflowBinary}
\alias{ReadModflowBinary}
\title{Read MODFLOW Binary File}
\usage{
ReadModflowBinary(path, data.type = c("array", "flow"), rm.totim.0 = FALSE)
}
\arguments{
\item{path}{character.
Path name of the binary file.}

\item{data.type}{character.
Description of how the data were saved.
Specify \code{"array"} for array data (such as hydraulic heads) and
\code{"flow"} for cell-by-cell flow data.}

\item{rm.totim.0}{logical.
If true, stress-periods at time zero are removed.}
}
\value{
Returns a \code{list} object of length equal to the
  number of times the data types are written to the binary file.
  The following list components are returned:
  \describe{
    \item{d}{matrix of values.
      The format typically coincides with a layer in the model grid.
      The one exception is for flow data (\code{data.type = "flow"}) in a
      cell-by-cell budget file saved using the \emph{"COMPACT BUDGET"} output option;
      for this case, matrix columns are cell index (\code{"icell"}),
      model-grid layer (\code{"layer"}), model-grid row (\code{"row"}),
      model-grid column (\code{"column"}), cell-by-cell flow (\code{"flow"}),
      and any auxiliary variables saved using the \emph{"AUXILIARY"} output option.}
    \item{kstp}{time step}
    \item{kper}{stress period}
    \item{desc}{description of data-type, such as "wells".}
    \item{ilay}{model-grid layer index}
    \item{delt}{length of the current time step.}
    \item{pertim}{time in the stress period.}
    \item{totim}{total elapsed time}
  }
  The layer index (ilay) and elapsed time (delt, pertim, totim) are only available
  when the cell-by-cell budget file is saved using the
  \emph{"COMPACT BUDGET"} output option.
}
\description{
This is a utility function for \href{https://water.usgs.gov/ogw/modflow/}{MODFLOW},
a U.S. Geological Survey groundwater-flow model.
It reads data from binary files produced by MODFLOW.
}
\details{
This function reads hydraulic head, drawdown,
  and cell budget files produced by MODFLOW.
}
\examples{
path <- system.file("extdata", "ex3D.hds", package = "inlmisc")
hds <- ReadModflowBinary(path, "array")
image(hds[[1]]$d)

}
\seealso{
\code{\link{SummariseBudget}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{IO}
