% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecreateLibrary.R
\name{RecreateLibrary}
\alias{RecreateLibrary}
\alias{SavePackageNames}
\title{Recreate R Library}
\usage{
RecreateLibrary(file = "package-names.txt", lib = NULL,
  repos = getOption("repos"))

SavePackageNames(file = "package-names.txt", lib = NULL)
}
\arguments{
\item{file}{'character'.
Name of the file for reading (or writing) the list of package names.
For file names that do not contain an absolute path,
the name is assumed relative to the current working directory, see \code{\link{getwd}()} command.}

\item{lib}{'character'.
The library tree(s) to search through when locating installed packages (see \code{\link{.libPaths}}),
or the library directory where to install packages.}

\item{repos}{'character'.
Vector of base URL(s) of the repositories to use when installing packages.
For example, the URL of the Geological Survey R Archive Network (GRAN) is \code{"https://owi.usgs.gov/R"}.}
}
\description{
These functions can be used to recreate an existing library on a new installation of \R.
}
\details{
A typical workflow is as follows:
Run the \code{SavePackageNames()} command on an older version of \R
to print to a text file a complete list of names for packages located under your current \R library tree(s).
If it is no longer needed, uninstall the older version of \R.
On a freshly installed version of \R, with the \pkg{inlmisc} package available,
run the \code{RecreateLibrary()} command to download and install
the packages listed in the package-names text file.
}
\examples{
# Run on old version of R
SavePackageNames()

\dontrun{
# Run on new version of R, and ensure inlmisc package is available.
repos <- c(CRAN = "https://cloud.r-project.org/", GRAN = "https://owi.usgs.gov/R")
if (!"inlmisc" \%in\% rownames(utils::installed.packages()))
  utils::install.packages("inlmisc", repos = repos["CRAN"])
inlmisc::RecreateLibrary(repos = repos)
}

}
\seealso{
\code{\link[utils]{installed.packages}}, \code{\link[utils]{install.packages}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
