% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToScientific.R
\name{ToScientific}
\alias{ToScientific}
\title{Format for Scientific Notation}
\usage{
ToScientific(x, digits = NULL, lab.type = c("latex", "plotmath"),
  na = as.character(NA), inline.delimiter = "$", scipen = NULL, ...)
}
\arguments{
\item{x}{'numeric'.
Vector of numbers}

\item{digits}{'integer'.
Number of digits after the decimal point for the mantissa.}

\item{lab.type}{'character'.
By default, LaTeX formatted strings for labels are returned.
Alternatively, \code{lab.type = "plotmath"} returns \code{\link[grDevices]{plotmath}}-compatible expressions.}

\item{na}{'character'.
String to be used for missing values (\code{NA}).
By default, no string substitution is made for missing values.}

\item{inline.delimiter}{'character'.
Delimiter for LaTeX inline mathematical mode.}

\item{scipen}{'integer'.
A penalty to be applied when deciding to format numeric values in scientific or fixed notation.
By default all numbers, with the exception of zero, are formatted using scientific notation.}

\item{...}{Arguments passed to the \code{\link{formatC}} function.
Only applies to fixed formatted values.}
}
\value{
For the default \code{lab.type = "latex"}, a 'character' vector of the same length as argument \code{x}.
  And for \code{lab.type = "plotmath"}, an expression of the same length as \code{x}.
  As a workaround for \href{https://www.section508.gov}{Section 508} compliance,
  the letter "x" is used as the times symbol in the plotmath-compatible expressions---rather
  than the more common (and better looking) "\%*\%" separator.
}
\description{
This function formats numbers in scientific notation \eqn{m \times 10^{n}}{m x 10^n}.
}
\examples{
x <- c(-1e+09, 0, NA, pi * 10^(-5:5))
ToScientific(x, digits = 5L, na = "---")
ToScientific(x, digits = 2L, scipen = 0L)

x <- exp(log(10) * 1:6)
i <- seq_along(x)
plot(i, i, type = "n", xaxt = "n", yaxt = "n", ann = FALSE)
lab <- ToScientific(x, digits = 0L, lab.type = "plotmath",
                    scipen = 0L, big.mark = ",")
axis(1, i, labels = lab)
axis(2, i, labels = lab)

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
