% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs_article.R
\name{usgs_article}
\alias{usgs_article}
\title{USGS Article Format}
\usage{
usgs_article(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown]{pdf_document}} function.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for creating a U.S. Geological Survey (USGS) article.
}
\examples{
\dontrun{
rmarkdown::draft("myarticle.Rmd", template = "usgs_article", package = "inlmisc")
rmarkdown::render("myarticle/myarticle.Rmd")
}

}
